move Core/IdeaColl/Proposal to Core/Proposal
