+++ src/adhocracy_frontend/adhocracy_frontend/static/js/Packages/Core/IdeaCollection/Poll/Poll.ts
@@ -3,9 +3,9 @@
 import * as AdhBadge from "../../Badge/Badge";
 import * as AdhConfig from "../../Config/Config";
 import * as AdhHttp from "../../Http/Http";
-import * as AdhIdeaCollectionProposal from "../Proposal/Proposal";
 import * as AdhPermissions from "../../Permissions/Permissions";
 import * as AdhProcess from "../../Process/Process";
+import * as AdhProposal from "../../Proposal/Proposal";
 import * as AdhRate from "../../Rate/Rate";
 import * as AdhTopLevelState from "../../TopLevelState/TopLevelState";
 
@@ -56,7 +56,7 @@ export var detailDirective = (
             if (!scope.processProperties && processType) {
                 scope.processProperties = adhProcess.getProperties(processType);
             }
-            AdhIdeaCollectionProposal.bindPath(adhConfig, adhHttp, adhPermissions, adhRate, adhTopLevelState, adhGetBadges, $q)(scope);
+            AdhProposal.bindPath(adhConfig, adhHttp, adhPermissions, adhRate, adhTopLevelState, adhGetBadges, $q)(scope);
 
             scope.goToLogin = () => {
                 adhTopLevelState.setCameFromAndGo("/login");
