+++ src/adhocracy_frontend/adhocracy_frontend/static/js/Packages/Core/Proposal/DetailSlot.html
