+++ src/adhocracy_frontend/adhocracy_frontend/static/js/Packages/Core/Proposal/Proposal.ts
@@ -1,32 +1,32 @@
-/// <reference path="../../../../../lib2/types/angular.d.ts"/>
+/// <reference path="../../../../lib2/types/angular.d.ts"/>
 
-import * as AdhBadge from "../../Badge/Badge";
-import * as AdhConfig from "../../Config/Config";
-import * as AdhHttp from "../../Http/Http";
-import * as AdhMapping from "../../Mapping/Mapping";
-import * as AdhPermissions from "../../Permissions/Permissions";
-import * as AdhPreliminaryNames from "../../PreliminaryNames/PreliminaryNames";
-import * as AdhProcess from "../../Process/Process";
-import * as AdhRate from "../../Rate/Rate";
-import * as AdhTopLevelState from "../../TopLevelState/TopLevelState";
-import * as AdhUtil from "../../Util/Util";
+import * as AdhBadge from "../Badge/Badge";
+import * as AdhConfig from "../Config/Config";
+import * as AdhHttp from "../Http/Http";
+import * as AdhMapping from "../Mapping/Mapping";
+import * as AdhPermissions from "../Permissions/Permissions";
+import * as AdhPreliminaryNames from "../PreliminaryNames/PreliminaryNames";
+import * as AdhProcess from "../Process/Process";
+import * as AdhRate from "../Rate/Rate";
+import * as AdhTopLevelState from "../TopLevelState/TopLevelState";
+import * as AdhUtil from "../Util/Util";
 
-import * as ResourcesBase from "../../../ResourcesBase";
+import * as ResourcesBase from "../../ResourcesBase";
 
-import RICommentVersion from "../../../../Resources_/adhocracy_core/resources/comment/ICommentVersion";
-import RISystemUser from "../../../../Resources_/adhocracy_core/resources/principal/ISystemUser";
-import * as SICommentable from "../../../../Resources_/adhocracy_core/sheets/comment/ICommentable";
-import * as SIDescription from "../../../../Resources_/adhocracy_core/sheets/description/IDescription";
-import * as SIImageReference from "../../../../Resources_/adhocracy_core/sheets/image/IImageReference";
-import * as SILocationReference from "../../../../Resources_/adhocracy_core/sheets/geo/ILocationReference";
-import * as SIMetadata from "../../../../Resources_/adhocracy_core/sheets/metadata/IMetadata";
-import * as SIMultiPolygon from "../../../../Resources_/adhocracy_core/sheets/geo/IMultiPolygon";
-import * as SIPoint from "../../../../Resources_/adhocracy_core/sheets/geo/IPoint";
-import * as SIRateable from "../../../../Resources_/adhocracy_core/sheets/rate/IRateable";
-import * as SITitle from "../../../../Resources_/adhocracy_core/sheets/title/ITitle";
-import * as SIVersionable from "../../../../Resources_/adhocracy_core/sheets/versions/IVersionable";
+import RICommentVersion from "../../../Resources_/adhocracy_core/resources/comment/ICommentVersion";
+import RISystemUser from "../../../Resources_/adhocracy_core/resources/principal/ISystemUser";
+import * as SICommentable from "../../../Resources_/adhocracy_core/sheets/comment/ICommentable";
+import * as SIDescription from "../../../Resources_/adhocracy_core/sheets/description/IDescription";
+import * as SIImageReference from "../../../Resources_/adhocracy_core/sheets/image/IImageReference";
+import * as SILocationReference from "../../../Resources_/adhocracy_core/sheets/geo/ILocationReference";
+import * as SIMetadata from "../../../Resources_/adhocracy_core/sheets/metadata/IMetadata";
+import * as SIMultiPolygon from "../../../Resources_/adhocracy_core/sheets/geo/IMultiPolygon";
+import * as SIPoint from "../../../Resources_/adhocracy_core/sheets/geo/IPoint";
+import * as SIRateable from "../../../Resources_/adhocracy_core/sheets/rate/IRateable";
+import * as SITitle from "../../../Resources_/adhocracy_core/sheets/title/ITitle";
+import * as SIVersionable from "../../../Resources_/adhocracy_core/sheets/versions/IVersionable";
 
-export var pkgLocation = "/Core/IdeaCollection/Proposal";
+export var pkgLocation = "/Core/Proposal";
 
 
 export interface IScope extends angular.IScope {
