+++ src/meinberlin_lib/meinberlin_lib/static/js/Packages/Meinberlin/Alexanderplatz/Workbench/Module.ts
@@ -1,7 +1,6 @@
 import * as AdhDocumentModule from "../../../Core/Document/Module";
 import * as AdhHttpModule from "../../../Core/Http/Module";
 import * as AdhMappingModule from "../../../Core/Mapping/Module";
-import * as AdhIdeaCollectionModule from "../../../Core/IdeaCollection/Module";
 import * as AdhMovingColumnsModule from "../../../Core/MovingColumns/Module";
 import * as AdhNamesModule from "../../../Core/Names/Module";
 import * as AdhPermissionsModule from "../../../Core/Permissions/Module";
@@ -9,9 +8,9 @@ import * as AdhProcessModule from "../../../Core/Process/Module";
 import * as AdhResourceAreaModule from "../../../Core/ResourceArea/Module";
 import * as AdhTopLevelStateModule from "../../../Core/TopLevelState/Module";
 
-import * as AdhIdeaCollectionProposal from "../../../Core/IdeaCollection/Proposal/Proposal";
 import * as AdhNames from "../../../Core/Names/Names";
 import * as AdhProcess from "../../../Core/Process/Process";
+import * as AdhProposal from "../../../Core/Proposal/Proposal";
 import * as AdhWorkbench from "../../../Core/Workbench/Workbench";
 
 import RIGeoProposal from "../../../../Resources_/adhocracy_core/resources/proposal/IGeoProposal";
@@ -30,7 +29,6 @@ export var register = (angular) => {
         .module(moduleName, [
             AdhDocumentModule.moduleName,
             AdhHttpModule.moduleName,
-            AdhIdeaCollectionModule.moduleName,
             AdhMappingModule.moduleName,
             AdhMovingColumnsModule.moduleName,
             AdhNamesModule.moduleName,
@@ -43,9 +41,9 @@ export var register = (angular) => {
             adhProcessProvider.templates[processType] = "<adh-meinberlin-alexanderplatz-workbench " +
                 "data-process-properties=\"processProperties\"></adh-meinberlin-alexanderplatz-workbench>";
             adhProcessProvider.setProperties(processType, {
-                createSlot: adhConfig.pkg_path + AdhIdeaCollectionProposal.pkgLocation + "/CreateSlot.html",
-                detailSlot: adhConfig.pkg_path + AdhIdeaCollectionProposal.pkgLocation + "/DetailSlot.html",
-                editSlot: adhConfig.pkg_path + AdhIdeaCollectionProposal.pkgLocation + "/EditSlot.html",
+                createSlot: adhConfig.pkg_path + AdhProposal.pkgLocation + "/CreateSlot.html",
+                detailSlot: adhConfig.pkg_path + AdhProposal.pkgLocation + "/DetailSlot.html",
+                editSlot: adhConfig.pkg_path + AdhProposal.pkgLocation + "/EditSlot.html",
                 hasAuthorInListItem: true,
                 hasCommentColumn: true,
                 hasDescription: true,
